Fixed up issues regarding dropsystem Fixed so you can now have multiple levels of clue scrolls active at one time.

+++ src/com/rs/kotlin/game/npc/drops/WeightedDropEntry.kt
@@ -1,4 +1,4 @@
 package com.rs.kotlin.game.npc.drops
 
-open class WeightedDropEntry(itemId: Int, minAmount: Int, maxAmount: Int, var weight: Int) :
+open class WeightedDropEntry(itemId: Int, minAmount: Int, maxAmount: Int, var weight: Double) :
     DropEntry(itemId, minAmount, maxAmount)
